Unit test isDescriptive() thoroughly and add special cases to dictionary JSON
It turns out that the dictionary JSON chosen for this task covers a lot more edge cases than expected. For one, it includes common abbreviations like "num", "msg", "temp", "min", and "max" that are often used as variable names (although I did have to manually add "tmp" to it, which is a common abbreviation for "temp"). Additionally, all words are given in lowercase, meaning that proper nouns can be properly validated even when we turn the upper case starting letters to lower case.